requirements.txt
openai>=1.1.0
anthropic>=0.9.0
requests>=2.28.0
python-dotenv>=1.0.0
tqdm>=4.65.0
importlib-metadata>=6.0.0
packaging>=23.0.0 